Refactor ActivityReplyView into ComposeView
UI changes between a "composing" view and a "replying" view depending on if the parent activity is null or not. Includes tests.
