+++ packages/stream_feed_flutter/lib/src/widgets/buttons/reply_button.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:stream_feed_flutter/src/theme/stream_feed_theme.dart';
-import 'package:stream_feed_flutter/src/widgets/dialogs/comment.dart';
 import 'package:stream_feed_flutter/src/widgets/icons.dart';
-import 'package:stream_feed_flutter/src/widgets/pages/activity_reply_view.dart';
+import 'package:stream_feed_flutter/src/widgets/pages/compose_view.dart';
 import 'package:stream_feed_flutter_core/stream_feed_flutter_core.dart';
 
 // ignore_for_file: cascade_invocations
@@ -59,7 +58,7 @@ class ReplyButton extends StatelessWidget {
           onPressed: () {
             Navigator.of(context).push(
               MaterialPageRoute(
-                builder: (_) => ActivityReplyView(
+                builder: (_) => ComposeView(
                   parentActivity: activity,
                   feedGroup: feedGroup,
                 ),
