+++ packages/stream_feed_flutter/lib/stream_feed_flutter.dart
@@ -14,7 +14,7 @@ export 'src/widgets/buttons/child_reaction.dart';
 export 'src/widgets/buttons/reaction.dart';
 export 'src/widgets/comment/field.dart';
 export 'src/widgets/icons.dart';
-export 'src/widgets/pages/activity_reply_view.dart';
+export 'src/widgets/pages/compose_view.dart';
 export 'src/widgets/pages/flat_feed_list_view.dart';
 export 'src/widgets/stream_feed_app.dart';
 export 'src/widgets/user/avatar.dart';
